WIP: dinput: Stop peeking messages where we shouldn't.
This causes DOSBOX to heavily stutter with high polling rate mouse,
for instance when running Doom II: Hell on Earth.

It was introduced by bfb845f9fccb2ff5bff0b0ba3238fec7f9f7b710, and the
comment indicates that it is supposed to fix Culpa Innata, which works
correctly without it, and Civilization: Call to Power 2, which doesn't
start in Proton anyway.